refactor(#91): 경로 수정
