+++ src/main/java/kr/co/souso/souso/global/enums/SortPageType.java
@@ -1,6 +1,6 @@
 package kr.co.souso.souso.global.enums;
 
-import kr.co.souso.souso.global.utils.code.Constant;
+import kr.co.souso.souso.global.utils.Constant;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
