+++ src/main/java/kr/co/souso/souso/global/utils/Constant.java
@@ -1,5 +1,7 @@
-package kr.co.souso.souso.global.utils.code;
+package kr.co.souso.souso.global.utils;
 
 public interface Constant {
+
     String getCode();
+
 }
